API documentation

Exported functions and types

CiteEXchange.Block — Type

A single block of CEX data.

CiteEXchange.cexversion — Function

Find CEX version for a block group.

cexversion(group)

Return nothing if no version specified.

CiteEXchange.blocks — Function

Parse a string into an Array of Blocks.

blocks(s)

CiteEXchange.blocktypes — Function

Find list of unique block types in a group.

blocktypes(blockgroup)

CiteEXchange.blocksfortype — Function

Find blocks of a given type in a blockgroup.

blocksfortype(blocktype, blockgroup)

CiteEXchange.datafortype — Function

Find data lines for all blocks of a given type in a blockgroup.

datafortype(blocktype, blockgroup)

CiteEXchange.relationsets — Function

Gather a (possibly empty) list of Cite2Urns identifying all relation sets in a list of Blocks.

relationsets(blocklist; delimiter)

If strict is true, all relations sets appearing in relationsets blocks must appear in a datamodel entry.

Internals

CiteEXchange.blocktype — Function

Determine block type from first line of a CEX block.

blocktype(s)