API documentation

Exported functions and types

CiteEXchange.Block — Type

A single block of CEX data.

CiteEXchange.cexversion — Function

Find CEX version for a block group.

cexversion(group)

Return nothing if no version specified.

CiteEXchange.blocks — Function

Parse a string into an Array of Blocks.

blocks(s)

CiteEXchange.blocktypes — Function

Find list of unique block types in a group.

blocktypes(blockgroup)

CiteEXchange.blocksfortype — Function

Find blocks of a given type in a blockgroup.

blocksfortype(blocktype, blockgroup)

CiteEXchange.datafortype — Function

Find blocks of a given type in a blockgroup.

datafortype(blocktype, blockgroup)

CiteEXchange.relations — Function

Extract data from all relation sets where relation belongs to a specified collection.

relations(blocklist, coll)

Internals

CiteEXchange.blocktype — Function

Determine block type from first line of a CEX block.

blocktype(s)